0.0.5 release

+++ scripts/code-gen.sh
@@ -2,6 +2,14 @@
 
 # required: git, [cargo-make](https://crates.io/crates/cargo-make/0.3.54#installation), [@cosmwasm/ts-codegen](https://www.npmjs.com/package/@cosmwasm/ts-codegen)
 
+TAG=$1
+
+if [ -z "$TAG" ]
+then
+  echo "TAG is empty"
+  exit 1
+fi
+
 rm -rf ./gmx_wasm
 rm -rf ./src/code-gen
 
@@ -11,7 +19,7 @@ git clone https://github.com/chadury2021/gmx_wasm.git
 
 cd gmx_wasm
 git fetch --tags
-git checkout v0.0.2
+git checkout $TAG
 
 cargo make generate-all-schemas
 cargo make ts-code-gen
